Streamline M2Crypto 3DES IV handling
M2Crypto defaults to padding the ciphertext,
and the previous implementation danced around that awkwardly
by padding and unpadding ciphertext on decryption
and updating the IV manually.
`m2.cipher_set_padding(context, 0)`
allows to shoulder the IV handling back to where it belongs
and to get rid of unnecessary context reinitializations.
(Investigative work courtesy of @tomato42:
 https://github.com/tomato42/tlslite-ng/pull/377#pullrequestreview-323893850)