Allow passing bytes to Python_AES
Python_AES makes a copy of plaintext and mutates it with item assignment.
Copying is currently done by slicing,
and this assumes a datatype that allows item assignment.
This commit replaces copying-by-slicing with copying by bytearray(...).
It also converts key and IV to bytearrays, so that passing an str works
on Python 2.