workaround HMAC behaviour in FIPS mode and deterministic ECDSA
The hmac module in python is actually FIPS compliant, it does that
by reverse-lookup of the hash function used as the parameter to the
initializer. Since normally we pass in a wrapped function (partial()),
it fails, and considers the function non FIPS approved.
By instead passing in the hash function, we can make it work
as long as the hash is FIPS compliant (so won't work for MD5).